refactor publishing & updating logic to reduce code duplication
